+++ src/widget/create_weight.rs
@@ -27,7 +27,7 @@ impl From<CreateWeightMessage> for Message {
 #[derive(Debug)]
 pub struct CreateWeight {
     day: DayFormField,
-    weight: InputFormField<f64>,
+    weight: InputFormField<f32>,
 }
 
 impl CreateWeight {
@@ -43,7 +43,7 @@ impl CreateWeight {
             if input.is_empty() {
                 Err(InputFormFieldError::MissingRequiredValue)
             } else {
-                match input.parse::<f64>() {
+                match input.parse::<f32>() {
                     Err(_) => Err(InputFormFieldError::InvalidNumber),
                     Ok(val) if val < 0.0 => Err(InputFormFieldError::SmallerThanZero),
                     Ok(val) => Ok(val),
