+++ src/widget/update_product.rs
@@ -32,10 +32,10 @@ pub struct UpdateProduct {
     product_id: usize,
     name: InputFormField<String>,
     company: InputFormField<Option<String>>,
-    calories: InputFormField<f64>,
-    fats: InputFormField<f64>,
-    proteins: InputFormField<f64>,
-    carbohydrates: InputFormField<f64>,
+    calories: InputFormField<f32>,
+    fats: InputFormField<f32>,
+    proteins: InputFormField<f32>,
+    carbohydrates: InputFormField<f32>,
 }
 
 impl UpdateProduct {
@@ -90,7 +90,7 @@ impl UpdateProduct {
             if input.is_empty() {
                 Err(InputFormFieldError::MissingRequiredValue)
             } else {
-                match input.parse::<f64>() {
+                match input.parse::<f32>() {
                     Err(_) => Err(InputFormFieldError::InvalidNumber),
                     Ok(val) if val < 0.0 => Err(InputFormFieldError::SmallerThanZero),
                     Ok(val) => Ok(val),
@@ -102,7 +102,7 @@ impl UpdateProduct {
             if input.is_empty() {
                 Err(InputFormFieldError::MissingRequiredValue)
             } else {
-                match input.parse::<f64>() {
+                match input.parse::<f32>() {
                     Err(_) => Err(InputFormFieldError::InvalidNumber),
                     Ok(val) if val < 0.0 => Err(InputFormFieldError::SmallerThanZero),
                     Ok(val) => Ok(val),
@@ -114,7 +114,7 @@ impl UpdateProduct {
             if input.is_empty() {
                 Err(InputFormFieldError::MissingRequiredValue)
             } else {
-                match input.parse::<f64>() {
+                match input.parse::<f32>() {
                     Err(_) => Err(InputFormFieldError::InvalidNumber),
                     Ok(val) if val < 0.0 => Err(InputFormFieldError::SmallerThanZero),
                     Ok(val) => Ok(val),
@@ -126,7 +126,7 @@ impl UpdateProduct {
             if input.is_empty() {
                 Err(InputFormFieldError::MissingRequiredValue)
             } else {
-                match input.parse::<f64>() {
+                match input.parse::<f32>() {
                     Err(_) => Err(InputFormFieldError::InvalidNumber),
                     Ok(val) if val < 0.0 => Err(InputFormFieldError::SmallerThanZero),
                     Ok(val) => Ok(val),
